Implement RegisterWidget flow with proceed and finished signals
Make SubManBaseWidget also a GObject.GObject. So that
sub classes of it can use gobject signals and properties.

Add 'proceed' signal to RegisterWidget, and a handler that
will move it to the next screen. To drive the widget, another
widget that provides cancel/register buttons makes them
emit the 'proceed' signal on registerWidget.

RegisterWidget also emits a 'finished' signal when registration
is done. The parent widget needs to deal with that. The default
RegisterDialog just hides the register window.