More use of signals/properties
Registration state (as used in global STATE) is now
a property on the registration widget, and the labels/titles
that need to be updated when it changes are hooked to the
notify signal on RegisterWidget's 'register_state' property.

Ditto for RegisterWidget's 'details-label-txt', ie, the
txt for the label under the progress bar ('Finding suitable service
levels', etc).

There is now a 'stay-on-screen' signal raise from applies or
async callbacks to indicate to stay on the current screen, replacing
the previous use of sending the widget to the 'DONT_CHANGE' screen.

'environment' is now tracked on the RegisterInfo class as a property.

RegisterWidget.pre_done is gone, and run_pre() will be shortly.