We don't need Screen.post() now, so remove it.
post() was used to take any values that had been
pulled out of the widgets, and update the shared
self._parent attributes, which was error prone.

Now apply() will update the properties on the shared
RegisterInfo instead, and if anything needs to follow
changes in those, it can connect to the notify signals.