wip, collapse error handling to one signal 'register-error'
Screen and NoGuiScreen, RegisterWidget, and RegisterDialog
all know of 'register-error' signal.

Screens can emit it, where RegisterWidget catches it, logs
it, stops emission of that signal. (Either the screens
or RegisterWidget could provide a class closure do_register_error
that that would see the signal first, and that can be used
to move to approriate screen to fix the error).
RegisterWidget then emits RegisterWidget's
'register-error' signal, that RegisterDialog is connected to.

RegisterDialog._on_register_error will call
RegisterDialogs handle_register_error or
handle_register_exception

The main smgui RegisterDialog just shows them in dialog windows,
but initial-setup will soon use it's set_error() method to show
the errors in place.

TODO: plug in more info to the register-error handlers args
to indicate severity of error, and potentially hints about
the register state at the time of the error (so the main window
dialog can decide if it wants to continue, or if it can be
ignored)