Split PerformRegisterScreen into two parts
One screen that creates the consumer, then do the
identity reloads etc in main thread, then do the
package profile upload in a new screen, in it's own
thread.

This gets consumer creation and profile upload out of
the same error handling block, so as to avoid package
profile upload failures causing a registration failure.

This also moves alot of the backend and asyncBackend
resetting and identity reloading back into the main
thread. We persist and reload the new consumer identity
after register now, before moving on to package profile
or sla selection. The facts updating, and installed
prod mgr caches still run in the register thread however.

Also, instead of calling restart_virt_who() deep in the
registration thread, now we wait until we persist/reload
the consumer identity, then emit a 'identity-updated' signal
on the RegisterInfo object. And at the moment, RegisterInfo
has a handler there that idle_add()'s restart_virt_who.
So we neither block on virt-who, or attempt to restart
it from within a thread.