Update rhel6 firstboot to use RegisterWidget
ditch the rhel5 firstboot compat stuff as well.

Some minor refactoring in registergui (add_screen)
that is likely not needed now.

TODO: the rhel6 firstboot wants a 'finish' screen
showing the 'pleast attach manually' message for example,
but we already hit the done screen, just need to set a
message for it.

the firstboot module is still kind of weird, and likely
should have always been a ModuleSet instead of a Module.
It wouldn't be hard to change now, but likely not worth it.