Add a ReloadConsumerScreen
This is a NonGuiScreen that is meant for handling the client
side post-registration tasks, including persisting the consumer
id cert and refreshing any objects that reference it.

This commit has the ActionClient call in find_service_levels
removed, but some variation of it needs to return for
cases where the RegisterWidget is used in AutoAttachDialog
mode. For the register case, most of ActionClient doesn't
make sense, as it either only applies to ent certs and contents,
or it's info that was just used for registration.

There is also a call to getConsumer in register that isn't
needed. We could also move the getServiceLevel list to happen
before auto attach and out of _find_service_levels (or in
find_service_levels but not _find_suitable_service_levels).