Use Notebook 'page' prop notify signal for tests
Instead of expecting the state of the notebook widget
to be updated by the time the 'register-error' handler
fires, instead wait for a notify signal from the notebooks
'page' property to indicate we successfully changes pages
(or that we didn't change pages in this case for
test_registergui.py)

Handle exceptions in gtk signal handlers better.

Keep track of any exceptions seen in self.exc_infos,
and after we end the gtk main loop, then see if we
captured any exceptions and raise them so nose fails
in a useful way.